Move TracePair and related functions out of the symbolic layer
