+++ doc/operators.rst
@@ -2,3 +2,4 @@ Discontinuous Galerkin Operators
 ================================
 
 .. automodule:: grudge.op
+.. automodule:: grudge.trace_pair
