+++ grudge/op.py
@@ -36,13 +36,18 @@ from pytools.obj_array import obj_array_vectorize, make_obj_array
 from meshmode.array_context import (
     FirstAxisIsElementsTag, make_loopy_program
 )
-from meshmode.dof_array import DOFArray, freeze, flatten, unflatten
-from meshmode.mesh import BTAG_ALL, BTAG_NONE, BTAG_PARTITION  # noqa
+from meshmode.dof_array import DOFArray, freeze
 
 import numpy as np
 import grudge.dof_desc as dof_desc
 
-from grudge.symbolic.primitives import TracePair
+from grudge.trace_pair import (  # noqa
+    interior_trace_pair,
+    interior_trace_pairs,
+    cross_rank_trace_pairs,
+    bdry_trace_pair,
+    bv_trace_pair
+)
 
 
 __doc__ = """
@@ -107,12 +112,6 @@ Elementwise reductions
 ----------------------
 
 .. autofunction:: elementwise_sum
-
-Distributed-memory functionality
---------------------------------
-
-.. autofunction:: interior_trace_pairs
-.. autofunction:: cross_rank_trace_pairs
 """
 
 
@@ -1122,170 +1121,6 @@ def elementwise_sum(dcoll, *args):
         actx,
         tuple(actx.call_loopy(prg(), operand=vec_i)["result"] for vec_i in vec)
     )
-
-# }}}
-
-
-# {{{ Interior trace pairs
-
-def _interior_trace_pair(dcoll, vec):
-    r"""Return a :class:`grudge.sym.TracePair` for the interior faces of
-    *dcoll* with a discretization tag specified by *discr_tag*.
-    This does not include interior faces on different MPI ranks.
-
-    :arg dcoll: a :class:`grudge.discretization.DiscretizationCollection`.
-    :arg vec: a :class:`~meshmode.dof_array.DOFArray` or object array of
-        :class:`~meshmode.dof_array.DOFArray`\ s.
-    :returns: a :class:`grudge.sym.TracePair` object.
-    """
-    i = project(dcoll, "vol", "int_faces", vec)
-
-    def get_opposite_face(el):
-        if isinstance(el, Number):
-            return el
-        else:
-            return dcoll.opposite_face_connection()(el)
-
-    e = obj_array_vectorize(get_opposite_face, i)
-
-    return TracePair("int_faces", interior=i, exterior=e)
-
-
-def interior_trace_pairs(dcoll, vec):
-    r"""Return a :class:`list` of :class:`grudge.sym.TracePair` objects defined
-    on the interior faces of *dcoll* and any faces connected to a parallel
-    boundary.
-
-    :arg dcoll: a :class:`grudge.discretization.DiscretizationCollection`.
-    :arg vec: a :class:`~meshmode.dof_array.DOFArray` or object array of
-        :class:`~meshmode.dof_array.DOFArray`\ s.
-    :returns: a :class:`list` of :class:`grudge.sym.TracePair` objects.
-    """
-    return (
-        [_interior_trace_pair(dcoll, vec)]
-        + cross_rank_trace_pairs(dcoll, vec)
-    )
-
-
-def interior_trace_pair(dcoll, vec):
-    from warnings import warn
-    warn("`interior_trace_pair` is deprecated and will be dropped "
-         "in version 2022.x. Use `interior_trace_pairs` instead, "
-         "which includes contributions from different MPI ranks.",
-         DeprecationWarning, stacklevel=2)
-    return _interior_trace_pair(dcoll, vec)
-
-# }}}
-
-
-# {{{ Distributed-memory functionality
-
-@memoize_on_first_arg
-def connected_ranks(dcoll):
-    from meshmode.distributed import get_connected_partitions
-    return get_connected_partitions(dcoll._volume_discr.mesh)
-
-
-class _RankBoundaryCommunication:
-    base_tag = 1273
-
-    def __init__(self, dcoll, remote_rank, vol_field, tag=None):
-        self.tag = self.base_tag
-        if tag is not None:
-            self.tag += tag
-
-        self.dcoll = dcoll
-        self.array_context = vol_field.array_context
-        self.remote_btag = BTAG_PARTITION(remote_rank)
-
-        self.bdry_discr = dcoll.discr_from_dd(self.remote_btag)
-        self.local_dof_array = project(dcoll, "vol", self.remote_btag, vol_field)
-
-        local_data = self.array_context.to_numpy(flatten(self.local_dof_array))
-
-        comm = self.dcoll.mpi_communicator
-
-        self.send_req = comm.Isend(
-                local_data, remote_rank, tag=self.tag)
-
-        self.remote_data_host = np.empty_like(local_data)
-        self.recv_req = comm.Irecv(self.remote_data_host, remote_rank, self.tag)
-
-    def finish(self):
-        self.recv_req.Wait()
-
-        actx = self.array_context
-        remote_dof_array = unflatten(self.array_context, self.bdry_discr,
-                actx.from_numpy(self.remote_data_host))
-
-        bdry_conn = self.dcoll.get_distributed_boundary_swap_connection(
-                dof_desc.as_dofdesc(dof_desc.DTAG_BOUNDARY(self.remote_btag)))
-        swapped_remote_dof_array = bdry_conn(remote_dof_array)
-
-        self.send_req.Wait()
-
-        return TracePair(self.remote_btag,
-                interior=self.local_dof_array,
-                exterior=swapped_remote_dof_array)
-
-
-def _cross_rank_trace_pairs_scalar_field(dcoll, vec, tag=None):
-    if isinstance(vec, Number):
-        return [TracePair(BTAG_PARTITION(remote_rank), interior=vec, exterior=vec)
-                for remote_rank in connected_ranks(dcoll)]
-    else:
-        rbcomms = [_RankBoundaryCommunication(dcoll, remote_rank, vec, tag=tag)
-                for remote_rank in connected_ranks(dcoll)]
-        return [rbcomm.finish() for rbcomm in rbcomms]
-
-
-def cross_rank_trace_pairs(dcoll, ary, tag=None):
-    r"""Get a :class:`list` of *ary* trace pairs for each partition boundary.
-
-    For each partition boundary, the field data values in *ary* are
-    communicated to/from the neighboring partition. Presumably, this
-    communication is MPI (but strictly speaking, may not be, and this
-    routine is agnostic to the underlying communication, see e.g.
-    _cross_rank_trace_pairs_scalar_field).
-
-    For each face on each partition boundary, a :class:`grudge.sym.TracePair`
-    is created with the locally, and remotely owned partition boundary face
-    data as the `internal`, and `external` components, respectively.
-    Each of the TracePair components are structured like *ary*.
-
-    :arg dcoll: a :class:`grudge.discretization.DiscretizationCollection`.
-    :arg ary: a single :class:`~meshmode.dof_array.DOFArray`, or an object
-        array of :class:`~meshmode.dof_array.DOFArray`\ s
-        of arbitrary shape.
-    :returns: a :class:`list` of :class:`grudge.sym.TracePair` objects.
-    """
-    if isinstance(ary, np.ndarray):
-        oshape = ary.shape
-        comm_vec = ary.flatten()
-
-        n, = comm_vec.shape
-        result = {}
-        # FIXME: Batch this communication rather than
-        # doing it in sequence.
-        for ivec in range(n):
-            for rank_tpair in _cross_rank_trace_pairs_scalar_field(
-                    dcoll, comm_vec[ivec]):
-                assert isinstance(rank_tpair.dd.domain_tag, dof_desc.DTAG_BOUNDARY)
-                assert isinstance(rank_tpair.dd.domain_tag.tag, BTAG_PARTITION)
-                result[rank_tpair.dd.domain_tag.tag.part_nr, ivec] = rank_tpair
-
-        return [
-            TracePair(
-                dd=dof_desc.as_dofdesc(
-                    dof_desc.DTAG_BOUNDARY(BTAG_PARTITION(remote_rank))),
-                interior=make_obj_array([
-                    result[remote_rank, i].int for i in range(n)]).reshape(oshape),
-                exterior=make_obj_array([
-                    result[remote_rank, i].ext for i in range(n)]).reshape(oshape)
-                )
-            for remote_rank in connected_ranks(dcoll)]
-    else:
-        return _cross_rank_trace_pairs_scalar_field(dcoll, ary, tag=tag)
 
 # }}}
